dpp-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
base58==2.1.1
bitarray==3.7.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
ckzg==2.1.2
click==8.2.1
cytoolz==1.0.1
deprecation==2.1.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipfshttpclient==0.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multiaddr==0.0.9
multidict==6.6.4
netaddr==1.3.0
packaging==25.0
parsimonious==0.10.0
pillow==11.3.0
postgrest==1.1.1
propcache==0.3.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyunormalize==16.0.0
qrcode==8.2
realtime==2.7.0
regex==2025.9.1
requests==2.32.5
rlp==4.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
toolz==1.0.0
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
varint==1.0.2
web3==7.13.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
